refactor: BattingManager 제거
